==> Codes/RCodes/00_scrape_census_data.Rmd <==
---
title: "Exploration of Neighborhood Racial Composition, Built Environment and Health Outcomes"
output: html_document
---

```{r message=FALSE}
library(tidycensus)
library(purrr)
library(dplyr)
library(tidyr)

knitr::opts_chunk$set(echo = TRUE)
```

Examples:  https://s3.amazonaws.com/assets.datacamp.com/production/course_6839/slides/chapter2.pdf
API Key: http://api.census.gov/data/key_signup.html
```{r}
# Set your API key
census_api_key("0d88f366fd4a202c99a944984292d25d8d7f00d6", install = TRUE, overwrite = TRUE)

# Check for variables 
variables_2019 <- load_variables(2019, "acs5", cache = TRUE)

# B01003_001 - Tota Population
# B01002_001 - Total Median Age
# B01001_026 - Total Female 
# B03002_003 - Non Hispanic White
# B03002_004 - Non Hispanic Blacks
# B03002_012 - Hispanic 
# B19013_001 - Median Household Income 
# B25064_001 - Median gross rent
# B25071_001 - Median gross rent as a percentage of household income
```

### 1. Race
```{r}
us <- unique(fips_codes$state)[1:51]

race_vars <- c(
  White = "B03002_003",
  Black = "B03002_004",
  Native = "B03002_005",
  Asian = "B03002_006",
  HIPI = "B03002_007",
  Hispanic = "B03002_012"
)

race_data <- map_df(us, function(x) {
  get_acs(geography = "tract", 
          year = 2019,
          variables = race_vars,
          summary_var = "B03002_001", # total population
          state = x)
})

saveRDS(race_data, file = "../../RawData/census/race_data.rds")


race_raw_data <- race_data %>%
  dplyr::select(GEOID, NAME, variable, estimate, summary_est) %>%
  pivot_wider(names_from = variable, values_from = estimate) %>%
  dplyr::rename(TotalPop = summary_est,
                Pop_White = White,
                Pop_Black = Black,
                Pop_Native = Native,
                Pop_Asian = Asian,
                Pop_HIPI = HIPI,
                Pop_Hispanic = Hispanic) 
saveRDS(race_raw_data, file = "../../RawData/census/race_raw_data.rds")

  
race_percent_data <- race_data %>%
  mutate(value = 100 * (estimate / summary_est)) %>%
  select(GEOID, NAME, variable, value)

# convert to wide form
race_percent_data <- race_percent_data %>% 
  pivot_wider(names_from = variable, values_from = value)

saveRDS(race_percent_data, file = "../../RawData/census/race_percent_data.rds")
```

### 2. Economic  
```{r}
us <- unique(fips_codes$state)[1:51]

econ_vars <- c(
  Median.gross.rent = "B25064_001",
  Median.household.income = "B19013_001",
  Rent.burden = "B25071_001"
)

econ_data <- map_df(us, function(x) {
  get_acs(geography = "tract", 
          year = 2019,
          variables = econ_vars,
          state = x)
})

econ_data <- econ_data %>%
  mutate(value = estimate) %>%
  select(GEOID, NAME, variable, value)

# convert to wide form
econ_data <- econ_data %>% 
  pivot_wider(names_from = variable, values_from = value)

saveRDS(econ_data, file = "../../RawData/census/econ_data.rds")
```

### 3. Education
```{r}
us <- unique(fips_codes$state)[1:51]

# Population over 25 with at least a bachelor’s degree is calculated as the sum of:
# Total Bachelor’s degree
# Total Master’s degree
# Total Professional degree
# Total Doctorate degree
# Proportion of population over 25 with at least a bachelor’s degree is calculated as 
# Population over 25 with at least a bachelor’s degree divded by the Total population over 25.
college_vars <- c("B15002_015",
                  "B15002_016",
                  "B15002_017",
                  "B15002_018",
                  "B15002_032",
                  "B15002_033",
                  "B15002_034",
                  "B15002_035")

college_data <- map_df(us, function(x) {
  get_acs(geography = "tract", 
          year = 2019,
          variables = college_vars,
          summary_var = "B15002_001", # total population over 25 
          state = x)
})

saveRDS(college_data, file = "../../RawData/census/college_data.rds")

# Calculate % 
college_percent_data <- college_data %>%
  group_by(GEOID, NAME) %>%
  dplyr::summarize(numerator = sum(estimate),
            denominator = first(summary_est)) %>%
  mutate(Pct_college = 100 * (numerator / denominator)) %>%
  dplyr::select(-numerator, -denominator)

saveRDS(college_percent_data, file = "../../RawData/census/college_percent_data.rds")
```

### 4. Age
```{r}
us <- unique(fips_codes$state)[1:51]

age_data <- map_df(us, function(x) {
  get_acs(geography = "tract", 
          year = 2019,
          variables = "B01002_001", # median age
          state = x)
})

age_data <- age_data %>%
  mutate(Age = estimate) %>%
  dplyr::select(GEOID, NAME, Age)

saveRDS(age_data, file = "../../RawData/census/age_data.rds")
```

### 5. Female
```{r}
us <- unique(fips_codes$state)[1:51]

female_data <- map_df(us, function(x) {
  get_acs(geography = "tract", 
          year = 2019,
          variables = "B01001_026",
          summary_var = "B01003_001", # total population
          state = x)
})

saveRDS(female_data, file = "../../RawData/census/female_data.rds")


female_percent_data <- female_data %>%
  mutate(value = 100 * (estimate / summary_est)) %>%
  select(GEOID, NAME, variable, value)

# convert to wide form
female_percent_data <- female_percent_data %>% 
  pivot_wider(names_from = variable, values_from = value) %>%
  dplyr::rename(Pct_Female = B01001_026)

saveRDS(female_percent_data, file = "../../RawData/census/female_percent_data.rds")
```

### 6. Insurance
```{r}
insurance_vars <- c("B27001_004",
                    "B27001_007",
                    "B27001_010",
                    "B27001_013",
                    "B27001_016",
                    "B27001_019",
                    "B27001_022",
                    "B27001_025",
                    "B27001_028",
                    "B27001_032",
                    "B27001_035",
                    "B27001_038",
                    "B27001_041",
                    "B27001_044",
                    "B27001_047",
                    "B27001_050",
                    "B27001_053",
                    "B27001_056")

insurance_data <- map_df(us, function(x) {
  get_acs(geography = "tract", 
          year = 2019,
          variables = insurance_vars,
          summary_var = "B27001_001", # total estimate
          state = x)
})

saveRDS(insurance_data, file = "../../RawData/census/insurance_data.rds")


# Calculate % 
insurance_percent_data <- insurance_data %>%
  group_by(GEOID, NAME) %>%
  dplyr::summarize(numerator = sum(estimate),
            denominator = first(summary_est)) %>%
  mutate(Pct_insurance = 100 * (numerator / denominator)) %>%
  dplyr::select(-numerator, -denominator)

saveRDS(insurance_percent_data, file = "../../RawData/census/insurance_percent_data.rds")
```

### 7. Owner-occupied housing
Neighborhood Stability: % living in home for 5 or more years, % owner occupied housing, % vacant housing
```{r}
us <- unique(fips_codes$state)[1:51]

owner_occupied_data <- map_df(us, function(x) {
  get_acs(geography = "tract", 
          year = 2019,
          variables = "B07013_002",
          summary_var = "B07013_001", # total 
          state = x)
})

saveRDS(owner_occupied_data, file = "../../RawData/census/owner_occupied_data.rds")


owner_occupied_percent_data <- owner_occupied_data %>%
  mutate(value = 100 * (estimate / summary_est)) %>%
  select(GEOID, NAME, variable, value)

# convert to wide form
owner_occupied_percent_data <- owner_occupied_percent_data %>% 
  pivot_wider(names_from = variable, values_from = value) %>%
  dplyr::rename(Pct_Owner = B07013_002)

saveRDS(owner_occupied_percent_data, file = "../../RawData/census/owner_occupied_percent_data.rds")
```

### 8. Vacant housing
```{r}
us <- unique(fips_codes$state)[1:51]

vacant_housing_data <- map_df(us, function(x) {
  get_acs(geography = "tract", 
          year = 2019,
          variables = "B25002_003",
          summary_var = "B25002_001", # total 
          state = x)
})

saveRDS(vacant_housing_data, file = "../../RawData/census/vacant_housing_data.rds")


vacant_housing_percent_data <- vacant_housing_data %>%
  mutate(value = 100 * (estimate / summary_est)) %>%
  select(GEOID, NAME, variable, value)

# convert to wide form
vacant_housing_percent_data <- vacant_housing_percent_data %>% 
  pivot_wider(names_from = variable, values_from = value) %>%
  dplyr::rename(Pct_Vacant = B25002_003)

saveRDS(vacant_housing_percent_data, file = "../../RawData/census/vacant_housing_percent_data.rds")
```

### 9. Single female head
```{r}
us <- unique(fips_codes$state)[1:51]

single_fam_data <- map_df(us, function(x) {
  get_acs(geography = "tract", 
          year = 2019,
          variables = "B09002_015", # female householder, no spouse present 
          summary_var = "B09002_001", # total 
          state = x)
})

saveRDS(single_fam_data, file = "../../RawData/census/single_fam_data.rds")


single_fam_percent_data <- single_fam_data %>%
  mutate(value = 100 * (estimate / summary_est)) %>%
  select(GEOID, NAME, variable, value)

# convert to wide form
single_fam_percent_data <- single_fam_percent_data %>% 
  pivot_wider(names_from = variable, values_from = value) %>%
  dplyr::rename(Pct_SingleFam = B09002_015)

saveRDS(single_fam_percent_data, file = "../../RawData/census/single_fam_percent_data.rds")
```
