==> Codes/RCodes/02_descriptive_statistics.Rmd <==
---
title: "Exploration of Neighborhood Racial Composition, Built Environment and Health Outcomes"
output: html_document
---

```{r message=FALSE}
library(sf)
library(tidyr)
library(tidyverse)
library(dplyr)
library(corrplot)
library(ggcorrplot)
library(reshape2)   
library(car)
library(reshape)
library(knitr)
library(stringr)
library(effects)
library(corrplot)
library(vtable)
library(metan)
library(ggplot2)
library(ggridges)
library(gtsummary)
knitr::opts_chunk$set(echo = TRUE)

# Load data 
df <- readRDS(file = "../../FinalData/finaldata.rds")
df_holc <- readRDS("../../FinalData/finaldata_holc.rds")
```

```{r, warning=FALSE, message = FALSE}
df$log.median.house.income  <- log(df$Median.household.income)
```

# Table 1: Descriptive statistics of health outcomes, built environment indicators, and census tract level covariates used in the model.
```{r}


# Tract Level 
df_holc  %>% 
  filter(Urban_Rural_New == 1) %>%
  dplyr::select(sleep_crud, diabetes_c, casthma_cr, prop_dilapbldg,  prop_green, prop_crosswalk, prop_not_single_family_home, prop_single_lane, Median.household.income, Pct_college, Age, Pct_Female, Pct_insurance, Pct_Owner, Pct_Vacant, Pct_SingleFam, majority_grade) %>% 
  tbl_summary(
    statistic = list(all_continuous() ~ "{mean} ({sd})",
                     all_categorical() ~ "{n} / {N} ({p}%)"),
    missing = "no",
    missing_text = 'Missing',
    digits = all_continuous() ~ 2,
    label = list(sleep_crud ~"Sleeping Problems", diabetes_c ~ "Diabetes",
                  casthma_cr~"Asthma", prop_dilapbldg~"% Dilapidated", prop_green~"% Green space", prop_crosswalk~"% Crosswalk", prop_not_single_family_home~"% Non-single Family Home", prop_single_lane~"% Single Lane", Median.household.income~"Median Household Income", Pct_college~"% College", Age~"Age", Pct_Female~"% Female", Pct_insurance~"% Insured", Pct_Owner~ "% Owner Occupied Housing", Pct_Vacant~"% Vacant Housing", Pct_SingleFam~"% Single Female Head of Household with Children", majority_grade~"HOLC Grade"
                 
                 )
  ) %>% add_n() 


# Population Level
df_holc %>% summarise(White = sum(Pop_White),
                      Black = sum(Pop_Black),
                       Native = sum(Pop_Native),
                      Asian = sum(Pop_Asian),
                      HIPI= sum(Pop_HIPI),
                      Hispanic= sum(Pop_Hispanic),
                      TotalPop=sum(TotalPop),
                     "% Non-Hispanic White" = sum(Pop_White)/sum(TotalPop),
                      "% Non-Hispanic Black" = sum(Pop_Black)/sum(TotalPop),
                      "% American Indian/Alaska Native" = sum(Pop_Native)/sum(TotalPop),
                      "% Asian" = sum(Pop_Asian)/sum(TotalPop),
                      "% Native Hawaiian and Other Pacific Islander"= sum(Pop_HIPI)/sum(TotalPop),
                      "% Hispanic"= sum(Pop_Hispanic)/sum(TotalPop),
                      ) %>% t()

```

# Figure S3: Distribution of built environment indicators and health outcomes, faceted by neighborhood topology. 
```{r}
df_composition <- df %>%
  mutate(Composition = ifelse(White >= 60 & Black < 20 & Hispanic < 20 & Asian < 20 & Native < 20 & HIPI < 20, "PW",
                       ifelse(Black >= 50 & White < 20 & Hispanic < 20 & Asian < 20 & Native < 20 & HIPI < 20, "PB",
                        ifelse(Black < 20 & White < 20 & (Asian >= 50 | Hispanic >= 50 | Native >= 50 | HIPI >= 50), "PNB", "Mixed"))))  %>%
  mutate(Composition = factor(Composition),
         bi_PW = ifelse(Composition == "PW", 1, 0),
         bi_PB = ifelse(Composition == "PB", 1, 0),
         bi_PNB = ifelse(Composition == "PNB", 1, 0),
         bi_mixed = ifelse(Composition == "Mixed", 1, 0)) %>%
  mutate(Pct_college = Pct_college / 100,
  Pct_Female = Pct_Female / 100,
  Pct_insurance = Pct_insurance / 100,
  Pct_Owner = Pct_Owner / 100, 
  Pct_Vacant = Pct_Vacant / 100,
  Pct_SingleFam = Pct_SingleFam /100,
  Urban_Rural_New = as.numeric(Urban_Rural_New))

df_composition$Composition <- relevel(df_composition$Composition, "PW")

df_composition <- df_composition %>% dplyr::select(-geometry) %>% 
  filter(Urban_Rural_New == 1)

be_cols=c("prop_dilapbldg", "prop_green",                "prop_crosswalk","prop_not_single_family_home", "prop_single_lane")

df_composition_melted <- df_composition %>% 
  dplyr::select(c(be_cols, 'Composition', diabetes_c,sleep_crud,casthma_cr)) %>% 
  mutate(
  diabetes_c=diabetes_c/100,
  sleep_crud=sleep_crud/100,
  casthma_cr=casthma_cr/100) %>% 
  pivot_longer(cols = -Composition)

ridge_plot=ggplot(df_composition_melted %>% drop_na(Composition), 
       aes(x = value, 
           y = factor(name, levels=c("prop_dilapbldg","prop_green","prop_crosswalk",
                                     "prop_not_single_family_home", "prop_single_lane",
                                     "diabetes_c","sleep_crud","casthma_cr")), 
           fill= stat(x))) +
  geom_density_ridges_gradient(scale = 3, rel_min_height = 0.01,panel_scaling=TRUE) +
  # scale_fill_viridis_c(name = "value", option = "pastel") +
  scale_fill_gradient2(low='#FFC4D0',mid = "#98CFD1", high='#E4CBF9',midpoint = 0.5,
                       name='Value')+
  scale_y_discrete(expand = c(0, 0), labels=c('% Dipalidated Building','% Green','% Crosswalk','% Not Single Family Homes','% Singel Lane',
                                              'Diabetes','Sleeping Problems','Asthma')) +
  scale_x_continuous(expand = c(0, 0)) +
  facet_wrap(.~factor(Composition, levels = c('PW','PB','PNB','Mixed'),
                      labels=c('Predominantly White',
                               'Predominantly Black',
                               'Predominantly minoritized racial group other than Black',
                               'Mixed')), scales = 'free_x') +
  theme_ridges()+theme(panel.spacing.x = unit(2, "lines"),
                       panel.spacing.y = unit(1, "lines"),
                       strip.text.x = element_text(size = 10,
                                                   margin = margin(b = 2, t = 2)),
                       axis.text = element_text(size = 8),
                       axis.title = element_blank(),
                       legend.text = element_text(size = 8),
                       legend.title = element_text(size = 10),
                       strip.background = element_rect(
     size=3.5
     ))
ridge_plot


ggsave('../../Figures/FigureS3.svg', ridge_plot, width = 11, height = 8)
```

# Figure S4: Correlation heatmap and Correlation Table with 95CI
```{r, warning=FALSE, message = FALSE}
# Create correlation plots
df_corr <- df %>%
  dplyr::filter(Urban_Rural_New == "1") %>%
  dplyr::select(-tract2010, -GEOID, -Census, -County, -State, -Urban_Rural_New) %>%
  dplyr::select(prop_dilapbldg, prop_green, 
                prop_crosswalk, prop_not_single_family_home, prop_single_lane, 
                White, Black, Asian, Hispanic,
                Median.household.income, Pct_college, Age, Pct_Female, 
                Pct_insurance, Pct_Owner, Pct_Vacant, Pct_SingleFam) %>%
  na.omit()

corr <- round(cor(df_corr), 2)
p.mat <- cor_pmat(df_corr)

corr_tab <-
  corr_ci(df_corr) %>% mutate(
    V1 = case_when(
      V1 == "sleep_crud" ~ "Sleeping Problems",
      V1 == "diabetes_c" ~ "Diabetes",
      V1 == "casthma_cr" ~ "Asthma",
      V1 == "prop_dilapbldg" ~ "% Dilapidated",
      V1 == "prop_green" ~ "% Green space",
      V1 == "prop_crosswalk" ~ "% Crosswalk",
      V1 == "prop_not_single_family_home" ~ "% Non-single Family Home",
      V1 == "prop_single_lane" ~ "% Single Lane",
      V1 == "Hispanic" ~ "% Hispanic",
      V1 == "Asian" ~ "% Non-Hispanic Asian",
      V1 == "Black" ~ "% Non-Hispanic Black",
      V1 == "White" ~ "% Non-Hispanic White",
      V1 == "Median.household.income" ~ "Median Household Income",
      V1 == "Pct_college" ~ "% College",
      V1 == "Age" ~ "Age",
      V1 == "Pct_Female" ~ "% Female",
      V1 == "Pct_insurance" ~ "% Insured",
      V1 == "Pct_Owner" ~ "% Owner Occupied Housing",
      V1 == "Pct_Vacant" ~ "% Vacant Housing",
      V1 =="Pct_SingleFam"~"% Single Female Head of Household with Children",
),
 V2 = case_when(
      V2 == "sleep_crud" ~ "Sleeping Problems",
      V2 == "diabetes_c" ~ "Diabetes",
      V2 == "casthma_cr" ~ "Asthma",
      V2 == "prop_dilapbldg" ~ "% Dilapidated",
      V2 == "prop_green" ~ "% Green space",
      V2 == "prop_crosswalk" ~ "% Crosswalk",
      V2 == "prop_not_single_family_home" ~ "% Non-single Family Home",
      V2 == "prop_single_lane" ~ "% Single Lane",
      V2 == "Hispanic" ~ "% Hispanic",
      V2 == "Asian" ~ "% Non-Hispanic Asian",
      V2 == "Black" ~ "% Non-Hispanic Black",
      V2 == "White" ~ "% Non-Hispanic White",
      V2 == "Median.household.income" ~ "Median Household Income",
      V2 == "Pct_college" ~ "% College",
      V2 == "Age" ~ "Age",
      V2 == "Pct_Female" ~ "% Female",
      V2 == "Pct_insurance" ~ "% Insured",
      V2 == "Pct_Owner" ~ "% Owner Occupied Housing",
      V2 == "Pct_Vacant" ~ "% Vacant Housing",
      V2 =="Pct_SingleFam"~"% Single Female Head of Household with Children",
))
sjPlot::tab_df(corr_tab,
               col.header=c('Variable 1','Variable 2','Corr','n','CI','LL',	'UL'),
               file = '../../Output/Supplementary/TableS1.html'
               ) 



corr.plot <- ggcorrplot(corr, 
                          p.mat = p.mat,
                          show.diag = T,
                          lab = T,
                          lab_size = 4, 
                          type = "lower",
                          outline.color = "white",
                          insig = "blank",
                          ggtheme = ggplot2::theme_minimal,
                          legend.title = "Correlation") +
  theme(axis.text = element_text(size = 7)) +
  scale_x_discrete(labels = c("% Dilapidated", "% Green space", "% Crosswalk",
                              "% Non-single family home", "% Single lane",
                              "% White", "% Black", "% Asian", "% Hispanic",
                              "Median Income", "% College", "Age", "% Female", 
                              "% Insured", "% Owner", "% Vacant", "% Single Family", "% Urbanity")) +
  scale_y_discrete(labels = c("% Dilapidated", "% Green space", "% Crosswalk",
                              "% Non-single family home", "% Single lane",
                              "% White", "% Black", "% Asian", "% Hispanic",
                              "Median Income", "% College", "Age", "% Female", 
                              "% Insured", "% Owner", "% Vacant", "% Single Family", "% Urbanity"))

corr.plot


ggsave('../../Figures/FigureS4.svg',corr.plot,width = 14, height =11 )
corr.plot
dev.off()
```
