==> Codes/RCodes/05_mediation_plots.Rmd <==
---
title: "Exploration of Neighborhood Racial Composition, Built Environment and Health Outcomes"
output: html_document
---

```{r message=FALSE}
library(dplyr)
library(tidyverse)
library(readxl)
library(ggplot2)

knitr::opts_chunk$set(echo = TRUE)

# Load data 
mediation <- read_excel("../../Output/Mediation.xlsx")
```

```{r}
mediation_total <- mediation %>% filter(Type == "Total Effect")

mediation_acme <- mediation %>% filter(Type == "ACME")

mediation_pct <- mediation %>% filter(Type == "Prop Mediated") %>%
  mutate(Estimate = Estimate * 100,
         LCI = LCI * 100,
         UCI = UCI * 100)

```

```{r}
figurea <- ggplot(mediation_acme %>% 
                    mutate(Race = fct_relevel(Race, "Mixed", "PNB", "PB")) %>%
                    mutate(Race = dplyr::recode(Race, 
                                                "PNB" = "White v.s. Non-Black",
                                                "PB" = "White v.s. Black",
                                                "Mixed" = "White v.s. Mixed"),
                           `Built Environment` = dplyr::recode(`Built Environment`,
                                        "prop_dilapbldg" = "Dilapidated Building",
                                        "prop_green" = "Green Space", 
                                        "prop_crosswalk" = "Crosswalks",
                                        "prop_not_single_family_home" = "Non Single Family Home",
                                        "prop_single_lane" = "Single Lane")),
                  aes(x=Race, y=Estimate)) + 
  geom_point(size = 0.4) + 
  geom_errorbar(width = 0.1, aes(ymin=LCI, ymax=UCI)) +
  facet_grid(fct_relevel(`Built Environment`,"Dilapidated Building", "Green Space", "Crosswalks", "Non Single Family Home", "Single Lane") ~ fct_relevel(Health, "Sleep", "Diabetes", "Asthma"), scales = "free_x") +
  scale_y_continuous(n.breaks = 4) +
  labs(subtitle = "(A) Average Causal Mediated Effect (ACME)") +
  xlab("") + ylab("") +
  coord_flip() +
  theme_light() +
  theme(strip.text.x = element_text(size = 10),
        strip.text.y = element_text(size = 10))
```

```{r}
figureb <- ggplot(mediation_pct %>% 
                    mutate(Race = fct_relevel(Race, "Mixed", "PNB", "PB")) %>%
                    mutate(Race = dplyr::recode(Race, 
                                                "PNB" = "White v.s. Non-Black",
                                                "PB" = "White v.s. Black",
                                                "Mixed" = "White v.s. Mixed"),
                           `Built Environment` = dplyr::recode(`Built Environment`,
                                        "prop_dilapbldg" = "Dilapidated Building",
                                        "prop_green" = "Green Space", 
                                        "prop_crosswalk" = "Crosswalks",
                                        "prop_not_single_family_home" = "Non Single Family Home",
                                        "prop_single_lane" = "Single Lane")),
                  aes(x=Race, y=Estimate)) + 
  geom_point(size = 0.4) + 
  geom_errorbar(width = 0.1, aes(ymin=LCI, ymax=UCI)) +
  facet_grid(fct_relevel(`Built Environment`,"Dilapidated Building", "Green Space", "Crosswalks", "Non Single Family Home", "Single Lane") ~ fct_relevel(Health, "Sleep", "Diabetes", "Asthma"), scales = "free_x") +
  scale_y_continuous(n.breaks = 4) +
  labs(subtitle = "(B) % Mediated") +
  xlab("") + ylab("") +
  coord_flip() +
  theme_light() +
  theme(strip.text.x = element_text(size = 10),
        strip.text.y = element_text(size = 10))
```

```{r}
library(ggpubr) 
figure <- ggpubr::ggarrange(figurea, figureb, ncol = 2)

jpeg(file = "../../Figures/FigureS5.jpeg", width = 15, height = 10, res = 300, units = 'in')
figure
dev.off()ß


jpeg(file = "../../Figures/FigureS5a.jpeg", width = 8.5, height =11, res = 300, units = 'in')
figurea
dev.off()


jpeg(file = "../../Figures/FigureS5b.jpeg", width = 8.5, height =11, res = 300, units = 'in')
figureb
dev.off()
```
